requirements.txt

Faker==37.3.0
psycopg2==2.9.10
robotframework==7.2.2
robotframework-appiumlibrary==2.1.0
robotframework-assertion-engine==3.0.3
robotframework-browser==19.5.1
robotframework-jsonlibrary==0.5
robotframework-requests==0.9.7
python_dotenv==1.1.0
